# WebApi/OdataWinClient/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EmployeeService",
    "version": "1.0"
  },
  "paths": {
...
    "/odata/CustomersApi/{id}": {
      "get": {
        "tags": [
          "CustomersApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;odata.metadata=minimal;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=minimal": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=minimal;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=minimal;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=full;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.metadata=none;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.streaming=true;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.streaming=true;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.streaming=false;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;odata.streaming=false;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;IEEE754Compatible=false": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/json;IEEE754Compatible=true": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CustomersApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json;odata.metadata=minimal;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=minimal": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/json;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "CustomersApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "contactTitle": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "fax": {
            "type": "string",
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Employee": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "titleOfCourtesy": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "hireDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "homePhone": {
            "type": "string",
            "nullable": true
          },
          "extension": {
            "type": "string",
            "nullable": true
          },
          "photo": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "reportsTo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "photoPath": {
            "type": "string",
            "nullable": true
          },
          "inverseReportsToNavigation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employee"
            },
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          },
          "reportsToNavigation": {
            "$ref": "#/components/schemas/Employee"
          }
        },
        "additionalProperties": false
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "employeeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "orderDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "requiredDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shippedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shipVia": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "freight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "shipName": {
            "type": "string",
            "nullable": true
          },
          "shipAddress": {
            "type": "string",
            "nullable": true
          },
          "shipCity": {
            "type": "string",
            "nullable": true
          },
          "shipRegion": {
            "type": "string",
            "nullable": true
          },
          "shipPostalCode": {
            "type": "string",
            "nullable": true
          },
          "shipCountry": {
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "employee": {
            "$ref": "#/components/schemas/Employee"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}